Reorganize Nailgun clients
