+++ appveyor.yml
@@ -21,4 +21,4 @@ build_script:
   - "mvn package"
 
 test_script:
-  - "python -m pynailgun.test_ng"
+  - "python -m nailgun-client.py.test_ng"
